# Conflicts:
#	src/server/scripts/Pandaria/zone_the_wandering_isle.cpp

+++ src/server/game/Server/Packets/CombatPackets.cpp
@@ -153,7 +153,7 @@ WorldPacket const* WorldPackets::Combat::PvPCredit::Write()
     _worldPacket << int32(OriginalHonor);
     _worldPacket << int32(Honor);
     _worldPacket << Target;
-    _worldPacket << int32(Rank);
+    _worldPacket << int8(Rank);
 
     return &_worldPacket;
 }
